+++ browser/templatingAPI.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var resolve = require('../lib/resolve');
+
+module.exports = {
+  resolve: resolve
+};